main/ingestion/requirements.txt
fastapi
re
uvicorn
requests
beautifulsoup4
sentence-transformers
qdrant-client
python-multipart
python-dotenv